POST hydration data setup
